+++ core/src/main/scala/com/krux/hyperion/HyperionCli.scala
@@ -18,8 +18,11 @@ trait HyperionCli { this: DataPipelineDef =>
     region: Option[String] = None,
     roleArn: Option[String] = None,
     output: Option[File] = None,
+    label: String = "id",
     removeLastNameSegment: Boolean = false,
     includeResources: Boolean = false,
+    includeDataNodes: Boolean = false,
+    includeDatabases: Boolean = false,
     tags: Map[String, Option[String]] = Map.empty
   )
 
@@ -35,8 +38,11 @@ trait HyperionCli { this: DataPipelineDef =>
       cmd("graph").action { (_, c) => c.copy(mode = "graph") }
         .children(
           opt[File]('o', "output").valueName("<file>").action { (x, c) => c.copy(output = Option(x)) },
+          opt[String]("label").valueName("id|name").action { (x, c) => c.copy(label = x) },
           opt[Unit]("remove-last-name-segment").action { (_, c) => c.copy(removeLastNameSegment = true) },
-          opt[Unit]("include-resources").action { (_, c) => c.copy(includeResources = true) }
+          opt[Unit]("include-resources").action { (_, c) => c.copy(includeResources = true) },
+          opt[Unit]("include-data-nodes").action { (_, c) => c.copy(includeDataNodes = true) },
+          opt[Unit]("include-databases").action { (_, c) => c.copy(includeDatabases = true) }
         )
 
       cmd("create").action { (_, c) => c.copy(mode = "create") }
@@ -80,7 +86,8 @@ trait HyperionCli { this: DataPipelineDef =>
           0
 
         case "graph" =>
-          val renderer = WorkflowGraphRenderer(this, cli.removeLastNameSegment, cli.includeResources)
+          val renderer = WorkflowGraphRenderer(this, cli.removeLastNameSegment, cli.label,
+            cli.includeResources, cli.includeDataNodes, cli.includeDatabases)
           cli.output.map(f => new PrintStream(f)).getOrElse(System.out).println(renderer.render())
           0