+++ core/src/main/scala/com/krux/hyperion/activity/MainClass.scala
@@ -4,6 +4,8 @@ import scala.language.implicitConversions
 
 class MainClass private (name: String) {
   override val toString = name
+  val fullName: String = name
+  val simpleName: String = name.split('.').last
 }
 
 object MainClass {
@@ -12,9 +14,9 @@ object MainClass {
   implicit def anyToMainClass(a: Any): MainClass = MainClass(a)
 
   def apply(mainClass: Any): MainClass = mainClass match {
-      case s: String => new MainClass(s.stripSuffix("$"))
-      case c: Class[_] => apply(c.getCanonicalName)
-      case mc => apply(mc.getClass)
+    case s: String => new MainClass(s.stripSuffix("$"))
+    case c: Class[_] => apply(c.getCanonicalName)
+    case mc => apply(mc.getClass)
   }
 }