+++ core/src/main/scala/com/krux/hyperion/workflow/WorkflowGraphRenderer.scala
@@ -11,7 +11,10 @@ import com.krux.hyperion.DataPipelineDef
 case class WorkflowGraphRenderer(
   pipeline: DataPipelineDef,
   removeLastNameSegment: Boolean,
-  includeResources: Boolean
+  label: String,
+  includeResources: Boolean,
+  includeDataNodes: Boolean,
+  includeDatabases: Boolean
 ) {
   private lazy val pipelineObjects: Seq[PipelineObject] = pipeline
 
@@ -19,20 +22,43 @@ case class WorkflowGraphRenderer(
     obj.getFields.asScala.find(_.getKey == "type").map(f => obj.getId -> f.getStringValue)
   }.toMap
 
-  private def quoted(s: String) = if (removeLastNameSegment) s""""${s.split('_').dropRight(1).mkString("_")}"""" else s""""$s""""
+  private lazy val idToLabelMap: Map[String, String] = pipelineObjects.flatMap { obj =>
+    label match {
+      case "id" => Option(obj.getId -> stripId(obj.getId))
+      case _ => obj.getFields.asScala.find(_.getKey == label).map(f => obj.getId -> Option(f.getStringValue).getOrElse(stripId(obj.getId)))
+    }
+  }.toMap
+
+  private def stripId(id: String) = if (removeLastNameSegment) id.split('_').dropRight(1).mkString("_") else id
+
+  private def quoted(s: String) = s""""$s""""
+
+  private def getLabel(id: String) = quoted(idToLabelMap.getOrElse(id, stripId(id)) match {
+    case "" => id
+    case idLabel => idLabel
+  })
 
   private def getAttributes(which: String, where: Config): Option[String] =
     Try(where.getObject(which)).toOption.map { conf =>
       "[" + conf.unwrapped().asScala.map { case (k, v) => s"$k=$v" }.mkString(", ") + "]"
     }
 
-  private def renderNode(id: String, attrs: String) = s"  ${quoted(id)} $attrs\n"
+  private def renderNode(id: String, attrs: String) = s"  ${getLabel(id)} $attrs\n"
 
   private def renderEdge(from: String, to: String) = {
     val attrs = getAttributes(s"${idToTypeMap(from)}To${idToTypeMap(to)}", pipeline.hc.graphStyles)
       .orElse(getAttributes(s"${idToTypeMap(from)}ToAny", pipeline.hc.graphStyles))
       .orElse(getAttributes(s"AnyTo${idToTypeMap(to)}", pipeline.hc.graphStyles))
-    s"  ${quoted(from)} -> ${quoted(to)} ${attrs.getOrElse("")}"
+    s"  ${getLabel(from)} -> ${getLabel(to)} ${attrs.getOrElse("")}"
+  }
+
+  private def ifNodeIncluded(nodeType: String)(func: => String): Option[String] = {
+    nodeType match {
+      case "Ec2Resource" | "EmrCluster" if !includeResources => None
+      case "DynamoDBDataNode" | "MySqlDataNode" | "RedshiftDataNode" | "S3DataNode" | "SqlDataNode" if !includeDataNodes => None
+      case "JdbcDatabase" | "RdsDatabase" | "RedshiftDatabase" if !includeDatabases => None
+      case _ => Option(func)
+    }
   }
 
   def render(): String = {
@@ -42,12 +68,19 @@ case class WorkflowGraphRenderer(
       obj.getFields.asScala.flatMap { field =>
         field.getKey match {
           case "type" =>
-            getAttributes(field.getStringValue, pipeline.hc.graphStyles).map(renderNode(obj.getId, _))
+            getAttributes(field.getStringValue, pipeline.hc.graphStyles).flatMap { attrs =>
+              ifNodeIncluded(field.getStringValue) {
+                renderNode(obj.getId, attrs)
+              }
+            }
 
-          case "output" =>
+          case "output" if includeDataNodes =>
             Option(renderEdge(obj.getId, field.getRefValue))
 
-          case "input" | "dependsOn" =>
+          case "input" if includeDataNodes =>
+            Option(renderEdge(field.getRefValue, obj.getId))
+
+          case "dependsOn" =>
             Option(renderEdge(field.getRefValue, obj.getId))
 
           case "runsOn" if includeResources =>